requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-telegram-bot==20.6
python-dotenv==1.0.0
gunicorn==21.2.0
aioflask==0.5.1
asgiref==3.7.2
